moved forestCharacteristics in a different file, added forestCharacteristics.primaryForest validation function
