+++ src/test/sectionSpec/fraYears.ts
@@ -1,0 +1,12 @@
+export const fraYears: Array<{ colName: string; cycles?: Array<string> }> = [
+  { colName: '1990' },
+  { colName: '2000' },
+  { colName: '2010' },
+  { colName: '2015' },
+  { colName: '2016', cycles: ['2020'] },
+  { colName: '2017', cycles: ['2020'] },
+  { colName: '2018', cycles: ['2020'] },
+  { colName: '2019', cycles: ['2020'] },
+  { colName: '2020' },
+  { colName: '2025', cycles: ['2025'] },
+]
